Add Linux installation info (#103)
* Add Linux installation info

Add a repology badge and mention the possibility to install via the package manager.

I believe this is a better approach than explicitly mentioning a repository, because it uses repology's database (and the list will remain updated).

* use tiny badge

The big badge is too big and could overflow in the following section. The tiny one will keep the same size as new packages are added.